Split out Payload.
